+++ Helper/Data.php
@@ -1,12 +1,16 @@
 <?php
 namespace Balancepay\Balancepay\Helper;
 
+use Balancepay\Balancepay\Model\Config as BalancepayConfig;
+use Balancepay\Balancepay\Model\Request\Factory as RequestFactory;
 use Balancepay\Balancepay\Model\ResourceModel\BalancepayProduct\CollectionFactory as MpProductCollection;
+use Magento\Customer\Api\CustomerRepositoryInterface;
+use Magento\Customer\Model\Session;
 use Magento\Framework\App\Cache\TypeListInterface;
-use Magento\Framework\App\Config\ReinitableConfigInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Framework\App\Http\Context;
 use Magento\Framework\Message\ManagerInterface as MessageManagerInterface;
+use Magento\Framework\Pricing\Helper\Data as PricingHelper;
 
 class Data extends AbstractHelper
 {
@@ -31,6 +35,31 @@ class Data extends AbstractHelper
     protected $appContext;
 
     /**
+     * @var Session
+     */
+    protected $customerSession;
+
+    /**
+     * @var CustomerRepositoryInterface
+     */
+    protected $customerRepositoryInterface;
+
+    /**
+     * @var RequestFactory
+     */
+    protected $requestFactory;
+
+    /**
+     * @var BalancepayConfig
+     */
+    protected $balancepayConfig;
+
+    /**
+     * @var PricingHelper
+     */
+    protected $pricingHelper;
+
+    /**
      * @param MpProductCollection $mpProductCollectionFactory
      * @param TypeListInterface $cacheTypeList
      * @param MessageManagerInterface $messageManager
@@ -40,12 +69,22 @@ class Data extends AbstractHelper
         MpProductCollection $mpProductCollectionFactory,
         TypeListInterface $cacheTypeList,
         MessageManagerInterface $messageManager,
-        Context $appContext
+        Context $appContext,
+        Session $customerSession,
+        CustomerRepositoryInterface $customerRepositoryInterface,
+        RequestFactory $requestFactory,
+        BalancepayConfig $balancepayConfig,
+        PricingHelper $pricingHelper
     ) {
         $this->_mpProductCollectionFactory = $mpProductCollectionFactory;
         $this->cacheTypeList = $cacheTypeList;
         $this->messageManager = $messageManager;
         $this->appContext = $appContext;
+        $this->customerSession = $customerSession;
+        $this->customerRepositoryInterface = $customerRepositoryInterface;
+        $this->requestFactory = $requestFactory;
+        $this->balancepayConfig = $balancepayConfig;
+        $this->pricingHelper = $pricingHelper;
     }
 
     /**
@@ -81,5 +120,55 @@ class Data extends AbstractHelper
     public function getCustomerSessionId()
     {
         return $this->appContext->getValue('customer_id');
+    }
+
+    public function getBuyerAmount()
+    {
+        $response = [];
+        try {
+            $buyerId = $this->getCustomerSessionBuyerId();
+            if (!empty($buyerId)) {
+                $response = $this->requestFactory
+                    ->create(RequestFactory::BUYER_REQUEST_METHOD)
+                    ->setRequestMethod('buyers/' . $buyerId)
+                    ->setTopic('getbuyers')
+                    ->process();
+            }
+        } catch (\Exception $e) {
+            $this->balancepayConfig->log('Get Buyer [Exception: ' .
+                $e->getMessage() . "]\n" . $e->getTraceAsString(), 'error');
+        }
+        return $response;
+    }
+
+    /**
+     * GetCustomerSessionBuyerId
+     *
+     * @return mixed
+     */
+    public function getCustomerSessionBuyerId()
+    {
+        if ($this->customerSession->getBuyerId()) {
+            return $this->customerSession->getBuyerId();
+        }
+        $customerId = $this->getCustomerSessionId();
+        if (!empty($customerId)) {
+            $customer = $this->customerRepositoryInterface->getById($customerId);
+            $customerAttributeData = $customer->__toArray();
+            $buyerId = isset($customerAttributeData['custom_attributes']['buyer_id']) ?
+                $customerAttributeData['custom_attributes']['buyer_id']['value'] : '';
+            $this->customerSession->setBuyerId($buyerId);
+            return $buyerId;
+        }
+        return 0;
+    }
+
+    /**
+     * @param $price
+     * @return float|string
+     */
+    public function formattedAmount($price)
+    {
+        return $this->pricingHelper->currency($price/100,true,false);
     }
 }
