+++ view/frontend/layout/customer_account.xml
@@ -1,0 +1,12 @@
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <body>
+        <referenceBlock name="customer_account_navigation">
+            <block class="Magento\Framework\View\Element\Html\Link\Current" name="customer-account-navigation-balancepay-payment-methods">
+                <arguments>
+                    <argument name="path" xsi:type="string">balancepay/payment/methods</argument>
+                    <argument name="label" xsi:type="string">Balance Payment Methods</argument>
+                </arguments>
+            </block>
+        </referenceBlock>
+    </body>
+</page>
